Refactor main.rb to handle only menu choices
